Removed member store
