--- src/store/members/getters.js
@@ -1,1 +1,0 @@
-export const members = ({ members }) => members
