--- src/store/members/mutations.js
@@ -1,7 +1,0 @@
-export const clearMembers = (state) => {
-  state.members = []
-}
-
-export const addMembers = (state, members) => {
-  state.members = [...state.members, ...members]
-}
