--- src/store/members/state.js
@@ -1,3 +1,0 @@
-export default () => ({
-  members: []
-})
